Streamlit_Audit_Migration/requirements.txt
streamlit
snowflake-connector-python
psycopg2-binary
python-dotenv
pandas
matplotlib
fpdf